Perform request to list endpoint

The low-level function to send one or more requests, depending on page, to the list endpoint.

Usage

bps_list(model, ..., domain_id = "0000", page = NULL, lang = c("ind", "eng"))

Arguments

model

The result model, e.g. "publication" to request a list of publications. See the Model section below for more details.

...

Name-value pairs that provide query parameters. See httr2::req_url_query() for more details.

domain_id

The domain ID of a BPS bureau. Defaults to "0000", which is the domain ID of the central bureau. Use bps_domain() to see the list of domain IDs.

page

The page number of the result. Defaults to page one. Set to Inf to request all pages.

lang

The result's language. Must be either "ind" for Indonesian or "eng" for English. Defaults to Indonesian.

Value

A list with the bpsr_table class.

Model

model must be one of the following:

"subject";

"subcat";

"var";

"vervar";

"turvar";

"turth";

"th";

"unit";

"news";

"newscategory";

"pressrelease";

"publication";

"statictable";

"subcatcsa";

"subjectcsa";

"tablestatistic";

"indicators";

"infographic";

"glosarium";

"sdgs";

"kbli2009";

"kbli2015";

"kbli2017";

"kbli2020"; or

"kbki2015".

The API actually also accepts "data" for model, but bpsr puts it in a separate family of functions due to differences in the response structure. See bps_get_dataset() and bps_get_datasets() for more details.

Examples

# Get a list of some publications
bps_list("publication", page = c(1, 2, 3), lang = "eng")
#> ℹ Performing paginated request...
#> ✔ Successfully requested 3 pages. [497ms]
#> 
#> <bpsr_table>
#> Table: publication
#> Content:
#> List of 2
#>  $ table      :List of 3
#>   ..$ :'data.frame':	10 obs. of  12 variables:
#>   .. ..$ pub_id   : chr [1:10] "fac1acc41ab7c4224bc09882" ...
#>   .. ..$ title    : chr [1:10] "Anti Corruption Behavior Index 2022" ...
#>   .. ..$ kat_no   : chr [1:10] "4407002" ...
#>   .. ..$ issn     : chr [1:10] "2622-8017" ...
#>   .. ..$ pub_no   : chr [1:10] "04300.2208" ...
#>   .. .. [list output truncated]
#>   ..$ :'data.frame':	10 obs. of  12 variables:
#>   .. ..$ pub_id   : chr [1:10] "70744a696d805364bdc1de8b" ...
#>   .. ..$ title    : chr [1:10] "Economic Indicators September 2022" ...
#>   .. ..$ kat_no   : chr [1:10] "9201001" ...
#>   .. ..$ issn     : chr [1:10] "" ...
#>   .. ..$ pub_no   : chr [1:10] "07300.2221" ...
#>   .. .. [list output truncated]
#>   ..$ :'data.frame':	10 obs. of  12 variables:
#>   .. ..$ pub_id   : chr [1:10] "0a09beac40484eaceb354a0b" ...
#>   .. ..$ title    : chr [1:10] "Directory of Rubber Plantation Establishment" ...
#>   .. ..$ kat_no   : chr [1:10] "1305075" ...
#>   .. ..$ issn     : chr [1:10] "2338-6983" ...
#>   .. ..$ pub_no   : chr [1:10] "05100.2214" ...
#>   .. .. [list output truncated]
#>  $ result_info:List of 3
#>   ..$ :List of 5
#>   .. ..$ page    : int 1
#>   .. ..$ pages   : int 481
#>   .. ..$ per_page: int 10
#>   .. ..$ count   : int 10
#>   .. ..$ total   : int 4803
#>   ..$ :List of 5
#>   .. ..$ page    : int 2
#>   .. ..$ pages   : int 481
#>   .. ..$ per_page: int 10
#>   .. ..$ count   : int 10
#>   .. ..$ total   : int 4803
#>   ..$ :List of 5
#>   .. ..$ page    : int 3
#>   .. ..$ pages   : int 481
#>   .. ..$ per_page: int 10
#>   .. ..$ count   : int 10
#>   .. ..$ total   : int 4803