Look up SDG dataset

Request the table of datasets related to the Sustainable Development Goals (SDG). This table contains dataset IDs, which are needed to request datasets using bps_get_dataset() or bps_get_datasets().

Usage

bps_sdg(goal = NULL, page = NULL, lang = c("ind", "eng"))

Arguments

goal

The goal number.

page

The page number of the result. Defaults to page one. Set to Inf to request all pages.

lang

The result's language. Must be either "ind" for Indonesian or "eng" for English. Defaults to Indonesian.

Value

A tibble containing dataset IDs and titles, among other things.

Examples

# Get a list of some datasets related to goal number one
bps_sdg(goal = 1, lang = "eng")
#> # A tibble: 10 × 12
#>    sdg_goal sdg_goal_…¹ sdg_id datas…² title subje…³ subject def   notes verti…⁴
#>    <chr>    <chr>       <chr>  <chr>   <chr> <chr>   <chr>   <chr> <chr> <chr>  
#>  1 sdgs_1   1. No Pove… 10.1.… 1238    Perc… 153     Geogra… NA    &lt;… 333    
#>  2 sdgs_1   1. No Pove… 1.2.1* 192     Perc… 23      Povert… NA    Data… 1      
#>  3 sdgs_1   1. No Pove… 1.2.1* 1539    Perc… 23      Povert… NA    &lt;… 396    
#>  4 sdgs_1   1. No Pove… 1.2.1* 1538    Perc… 23      Povert… NA    &lt;… 217    
#>  5 sdgs_1   1. No Pove… 1.2.1* 184     Perc… 23      Povert… NA    &lt;… 99     
#>  6 sdgs_1   1. No Pove… 1.4.1  2016    Prop… 12      Popula… NA    NA    1      
#>  7 sdgs_1   1. No Pove… 1.4.1  2017    Prop… 12      Popula… NA    NA    310    
#>  8 sdgs_1   1. No Pove… 1.4.2  2020    Prop… 29      Housing NA    &lt;… 301    
#>  9 sdgs_1   1. No Pove… 1.4.2  2019    Prop… 29      Housing NA    &lt;… 310    
#> 10 sdgs_1   1. No Pove… 1.4.2  2018    Prop… 29      Housing NA    &lt;… 1      
#> # … with 2 more variables: unit <chr>, graph <int>, and abbreviated variable
#> #   names ¹​sdg_goal_name, ²​dataset_id, ³​subject_id, ⁴​vertical_var_group_id