Look up subject category and subject of dataset

Get tables of dataset subject category and subject. The subject table contains subject IDs, which are useful to request specific records in the dataset table using bps_dataset().

bps_subject_category() requests the subject category table

bps_subject() requests the subject table

Usage

bps_subject_category(domain_id = "0000", page = NULL, lang = c("ind", "eng"))

bps_subject(
  subject_category_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

Arguments

domain_id

The domain ID of a BPS bureau. Defaults to "0000", which is the domain ID of the central bureau. Use bps_domain() to see the list of domain IDs.

page

The page number of the result. Defaults to page one. Set to Inf to request all pages.

lang

The result's language. Must be either "ind" for Indonesian or "eng" for English. Defaults to Indonesian.

subject_category_id

The subject category ID. Use bps_subject_category() to see the list of subject category IDs.

Value

bps_subject_category() returns a tibble containing subject category IDs and titles.

bps_subject() returns a tibble containing subject IDs and titles, as well as subject categories.

Examples

# Look up the subject category
bps_subject_category(lang = "eng")
#> # A tibble: 4 × 2
#>   subject_category_id title                 
#>   <chr>               <chr>                 
#> 1 0                   Other                 
#> 2 1                   Social and Population 
#> 3 2                   Economic and Trade    
#> 4 3                   Agriculture and Mining

# Look up the subjects in the economic and trade category
bps_subject(subject_category_id = "2", lang = "eng")
#> # A tibble: 10 × 5
#>    subject_id title                                subject_cate…¹ subje…² ntabel
#>    <chr>      <chr>                                <chr>          <chr>   <lgl> 
#>  1 25         Business & Consumer Tendency Index   2              Econom… NA    
#>  2 2          Communication                        2              Econom… NA    
#>  3 4          Construction                         2              Econom… NA    
#>  4 3          Consumer Prices Indices              2              Econom… NA    
#>  5 173        Domestic Trade                       2              Econom… NA    
#>  6 7          Energy                               2              Econom… NA    
#>  7 22         Farmer Terms of Trade                2              Econom… NA    
#>  8 100        Flow of Funds                        2              Econom… NA    
#>  9 8          Foreign Trade                        2              Econom… NA    
#> 10 169        Gross Domestic Product (Expenditure) 2              Econom… NA    
#> # … with abbreviated variable names ¹​subject_category_id, ²​subject_category