Look up dataset variable

Get tables of dataset variables. Each table contains the variable IDs, which are useful to request specific observations in a dataset using bps_get_dataset().

bps_vertical_var() requests the vertical variable table

bps_derived_var() requests the derived variable table

bps_year() requests the year table

bps_period() requests the period table

Usage

bps_vertical_var(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_derived_var(
  dataset_id = NULL,
  derived_var_group_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_year(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_period(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

Arguments

dataset_id

The dataset ID. Use bps_dataset() to see the list of dataset IDs.

domain_id

The domain ID of a BPS bureau. Defaults to "0000", which is the domain ID of the central bureau. Use bps_domain() to see the list of domain IDs.

page

The page number of the result. Defaults to page one. Set to Inf to request all pages.

lang

The result's language. Must be either "ind" for Indonesian or "eng" for English. Defaults to Indonesian.

derived_var_group_id

The derived variable group ID. Use bps_derived_var() to see the list of derived variable group IDs.

Value

bps_vertical_var() returns a tibble containing vertical variable IDs, names and groups.

bps_derived_var() returns a tibble containing derived variable IDs, names and groups.

bps_year() returns a tibble containing year IDs and years.

bps_period() returns a tibble containing period IDs and periods.

Examples

# Look up the vertical variables in the gross domestic product dataset
bps_vertical_var(dataset_id = "1955", lang = "eng")
#> # A tibble: 10 × 5
#>    vertical_var_id vertical_var                          verti…¹ verti…² verti…³
#>    <chr>           <chr>                                 <chr>   <chr>   <chr>  
#>  1 100             1. Household Consumption Expenditures 1630    91      GDP by…
#>  2 110             a. Food and Beverages, Other Than Re… 1631    91      GDP by…
#>  3 120             b. Clothing, Footwear and Maintenanc… 1632    91      GDP by…
#>  4 130             c. Household Housing and Equipment    1633    91      GDP by…
#>  5 140             d. Health and Education               1634    91      GDP by…
#>  6 150             e. Transportation and Communication   1635    91      GDP by…
#>  7 160             f. Restaurants and Hotels             1636    91      GDP by…
#>  8 170             g. Others                             1637    91      GDP by…
#>  9 200             2. Consumption Expenditures of Non-P… 1638    91      GDP by…
#> 10 300             3. Government Consumption Expenditur… 1639    91      GDP by…
#> # … with abbreviated variable names ¹​vertical_var_item_id,
#> #   ²​vertical_var_group_id, ³​vertical_var_group